Add base view class
