+++ lib/views/qiita-section-view.js
@@ -5,14 +5,10 @@
 import { File } from 'atom'
 import etch from 'etch'
 import fs from 'fs-plus'
+import SectionViewBase from './section-view-base'
 import QiitaService from '../services/qiita-service'
 
-export default class QiitaSectionView {
-  constructor(props) {
-    this.props = props || {}
-    etch.initialize(this)
-  }
-
+export default class QiitaSectionView extends SectionViewBase {
   render() {
     return (
       <section className="qiita">
@@ -59,21 +55,6 @@ export default class QiitaSectionView {
         </button>
       </section>
     )
-  }
-
-  update(newProps) {
-    Object.entries(newProps).forEach(([key, newValue]) => {
-      this.props[key] = newValue
-    })
-    return etch.update(this)
-  }
-
-  async destroy() {
-    await etch.destroy(this)
-  }
-
-  getElement() {
-    return this.element
   }
 
   importFromQiita(event) {
