+++ lib/views/section-view-base.js
@@ -1,0 +1,12 @@
+'use babel'
+
+/** @jsx etch.dom */
+
+import { File } from 'atom'
+import etch from 'etch'
+import fs from 'fs-plus'
+import ViewBase from './view-base'
+
+// TODO: - common section implementations.
+export default class SectionViewBase extends ViewBase {
+}
