+++ lib/views/sync-sync-view.js
@@ -6,13 +6,9 @@ import { File } from 'atom'
 import etch from 'etch'
 import fs from 'fs-plus'
 import QiitaSectionView from './qiita-section-view'
+import ViewBase from './view-base'
 
-export default class SyncSyncView {
-  constructor(props) {
-    this.props = props || {}
-    etch.initialize(this)
-  }
-
+export default class SyncSyncView extends ViewBase {
   render() {
     return (
       <div className="sync-sync native-key-bindings">
@@ -64,20 +60,5 @@ export default class SyncSyncView {
         </form>
       </div>
     )
-  }
-
-  update(newProps) {
-    Object.entries(newProps).forEach(([key, newValue]) => {
-      this.props[key] = newValue
-    })
-    return etch.update(this)
-  }
-
-  async destroy() {
-    await etch.destroy(this)
-  }
-
-  getElement() {
-    return this.element
   }
 }
