+++ lib/views/view-base.js
@@ -1,0 +1,29 @@
+'use babel'
+
+/** @jsx etch.dom */
+
+import { File } from 'atom'
+import etch from 'etch'
+import fs from 'fs-plus'
+
+export default class ViewBase {
+  constructor(props) {
+    this.props = props || {}
+    etch.initialize(this)
+  }
+
+  update(newProps) {
+    Object.entries(newProps).forEach(([key, newValue]) => {
+      this.props[key] = newValue
+    })
+    return etch.update(this)
+  }
+
+  async destroy() {
+    await etch.destroy(this)
+  }
+
+  getElement() {
+    return this.element
+  }
+}
